happy-solutions-crm/backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
pymongo==4.8.0
bcrypt==4.2.0
python-dotenv==1.0.1
requests==2.32.3
reportlab==4.2.2
